PCIE Unified device manager fat32 support

+++ kernel/src/drivers/pci/pci.h
@@ -25,12 +25,12 @@ struct pci_device {
 
 // Read/write PCI configuration space
 uint32_t pci_config_read32(uint8_t bus, uint8_t slot, uint8_t func,
-                           uint8_t offset);
-void pci_config_write32(uint8_t bus, uint8_t slot, uint8_t func, uint8_t offset,
+                           uint16_t offset);
+void pci_config_write32(uint8_t bus, uint8_t slot, uint8_t func, uint16_t offset,
                         uint32_t value);
 uint16_t pci_config_read16(uint8_t bus, uint8_t slot, uint8_t func,
-                           uint8_t offset);
-void pci_config_write16(uint8_t bus, uint8_t slot, uint8_t func, uint8_t offset,
+                           uint16_t offset);
+void pci_config_write16(uint8_t bus, uint8_t slot, uint8_t func, uint16_t offset,
                         uint16_t value);
 
 // Initialize PCI and enumerate all devices
